Refactor: Split init.lua into smaller modules
I've decomposed the main `init.lua` file into several smaller, more focused modules:
- `config.lua`: Handles configuration loading and defaults.
- `api.lua`: Manages communication with the Code::Stats API, including request sending and API-specific error handling.
- `events.lua`: Sets up Neovim autocommands for tracking activity.

The main `init.lua` file now acts as an orchestrator, requiring these modules and exposing the original public API.

This refactoring improves code organization and maintainability without changing existing functionality. Error handling has also been made more explicit within the `api` module.
